requirements.txt
python-dotenv
pyserial
matplotlib
numpy